extract TryGetExpressionComparedToNull as extension method
